% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfsdgsgs.R
\name{make_dwarf_age}
\alias{make_dwarf_age}
\title{Randomly select dwarf age from a range}
\usage{
make_dwarf_age(min_age, max_age)
}
\arguments{
\item{min_age}{Minimum age of the dwarf.}

\item{max_age}{Maximum age of the dwarf.}
}
\value{
A single integer with the age of the dwarf.
}
\description{
Randomly select dwarf age from a range
}
