% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfsdgsgs.R
\name{make_dwarf_name}
\alias{make_dwarf_name}
\title{Function for making silly dwarf names}
\usage{
make_dwarf_name(make_more_metal = TRUE)
}
\arguments{
\item{make_more_metal}{Should certain vowels be replaced with umlauts? Default is \code{TRUE}.}
}
\value{
A single dwarf name
}
\description{
Function for making silly dwarf names
}
