% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dwarf.R
\name{print_dwarf}
\alias{print_dwarf}
\title{Print a random dwarf character with name, age, and job.}
\usage{
print_dwarf(min_age = 18, max_age = 150, make_more_metal = T)
}
\arguments{
\item{min_age}{(Optional) The minimum age of the dwarf (numeric)}

\item{max_age}{(Optional) The maximum age of the dwarf (numeric)}

\item{make_more_metal}{Should certain vowels be replaced with umlauts? Default is \code{TRUE}.}
}
\value{
A single string with the dwarf's name, age, and job.
}
\description{
Print a random dwarf character with name, age, and job.
}
\examples{
print_dwarf(min_age=30,max_age=60)
}
