% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfsdgsgs.R
\name{make_dwarf_name}
\alias{make_dwarf_name}
\title{Function for making silly dwarf names}
\usage{
make_dwarf_name()
}
\value{
A dwarf name
}
\description{
Function for making silly dwarf names
}
\examples{
make_dwarf_name()

}
